Made SidebarCount a connected component
