+++ src/components/Sidebar/SidebarCount.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { connect } from "react-redux";
 import { BlockPadded } from "../styled-components/Block";
 import {
   SidebarCountParent,
@@ -6,12 +7,28 @@ import {
   SidebarCountLabel,
   SidebarCountValue,
 } from "../styled-components/SidebarCountStyles";
+import { getDiseaseCount } from "../../utils/sidebarDataHelpers";
+import { getFutureProjectionCount } from "../../utils/ebolaDataHelpers";
 
 const SidebarCount = ({
   filters,
-  diseaseCount = 0,
-  projectedDiseaseCount = 0,
+  ebolaData,
+  ebolaDataCombined,
+  covidCaseCountData,
+  covidDeathCountData,
 }) => {
+  const diseaseCount = getDiseaseCount(
+    ebolaData,
+    covidCaseCountData,
+    covidDeathCountData,
+    filters
+  );
+
+  const projectedDiseaseCount = getFutureProjectionCount(
+    ebolaData,
+    ebolaDataCombined,
+    filters
+  );
   // If the Covid outbreak is selected, display either "cases" or "deaths" depending which data type is selected.
   // If the Ebola outbreak is selected, just display "cases".
   const dataType = filters.outbreak === "COVID 19" ? filters.dataType : "cases";
@@ -37,17 +54,27 @@ const SidebarCount = ({
       <SidebarCountParent>
         <SidebarCountLabel>{labelText}</SidebarCountLabel>
         <SidebarCountColor style={{ backgroundColor: iconColor }} />
-        <SidebarCountValue>{diseaseCount}</SidebarCountValue>
+        <SidebarCountValue>{diseaseCount.toLocaleString()}</SidebarCountValue>
       </SidebarCountParent>
       {filters.dataType.includes("projected") && (
         <SidebarCountParent>
           <SidebarCountLabel>Projected future cases</SidebarCountLabel>
           <SidebarCountColor style={{ backgroundColor: "#F2AD33" }} />
-          <SidebarCountValue>{projectedDiseaseCount}</SidebarCountValue>
+          <SidebarCountValue>
+            {projectedDiseaseCount.toLocaleString()}
+          </SidebarCountValue>
         </SidebarCountParent>
       )}
     </BlockPadded>
   );
 };
 
-export default SidebarCount;
+const mapStateToProps = (state) => ({
+  filters: state.filters,
+  ebolaData: state.ebola.ebolaData.data,
+  ebolaDataCombined: state.ebola.ebolaDataCombined.data,
+  covidCaseCountData: state.covid.caseCounts.data,
+  covidDeathCountData: state.covid.deathCounts.data,
+});
+
+export default connect(mapStateToProps)(SidebarCount);
