+++ src/components/Sidebar/__tests__/SidebarCount.test.js
@@ -1,12 +1,28 @@
 import React from "react";
-import { shallow } from "enzyme";
+import { Provider } from "react-redux";
+import configureStore from "redux-mock-store";
+import thunk from "redux-thunk";
 import SidebarCount from "../SidebarCount";
+import renderer from "react-test-renderer";
+import "jest-styled-components";
 import { reduxInitialState } from "../../../constants/CommonTestData";
 
-const filtersState = reduxInitialState.filters;
+const mockStore = configureStore([thunk]);
 
-describe("Tests for the SidebarCount component", () => {
-  test("renders SidebarCount component", () => {
-    shallow(<SidebarCount filters={filtersState} diseaseCaseCount={5000} />);
+describe("Tests for the connected SidebarCount component with reduxInitialState", () => {
+  let store;
+  let component;
+
+  beforeEach(() => {
+    store = mockStore(reduxInitialState);
+
+    component = renderer.create(
+      <Provider store={store}>
+        <SidebarCount />
+      </Provider>
+    );
+  });
+  test("should render with given state from Redux store", () => {
+    expect(component.toJSON()).toMatchSnapshot();
   });
 });
