+++ src/components/Sidebar/index.js
@@ -21,17 +21,11 @@ import {
 } from "../styled-components/SelectWrappers";
 import { Button } from "../styled-components/Button";
 import { InputLabel } from "../styled-components/InputLabel";
-import { getFutureProjectionCount } from "../../utils/ebolaDataHelpers";
-import { getDiseaseCount } from "../../utils/sidebarDataHelpers";
 import CountrySelect from "../CountrySelect";
 import DataRadioButtons from "../DataRadioButtons";
 
 const Sidebar = ({
   filters,
-  ebolaData,
-  ebolaDataCombined,
-  covidCaseCountData,
-  covidDeathCountData,
   changeCountryFilter,
   changeOutbreakFilter,
   openDateRangePopover,
@@ -56,21 +50,6 @@ const Sidebar = ({
     // Open the date range popover.
     openDateRangePopover();
   };
-
-  // This is the disease count for the SidebarCount child component
-  const diseaseCount = getDiseaseCount(
-    ebolaData,
-    covidCaseCountData,
-    covidDeathCountData,
-    filters
-  );
-
-  // This is the projected disease count for the SidebarCount child component
-  const projectedDiseaseCount = getFutureProjectionCount(
-    ebolaData,
-    ebolaDataCombined,
-    filters
-  );
 
   const showSidebarCount = filters.view === "snapshot";
   const showEbolaRiskList =
@@ -99,13 +78,7 @@ const Sidebar = ({
         </Button>
       </div>
       <DataRadioButtons />
-      {showSidebarCount && (
-        <SidebarCount
-          filters={filters}
-          diseaseCount={diseaseCount.toLocaleString()}
-          projectedDiseaseCount={projectedDiseaseCount.toLocaleString()}
-        />
-      )}
+      {showSidebarCount && <SidebarCount />}
       {showEbolaRiskList && <EbolaRiskList />}
     </Styled.SidebarWrapper>
   );
@@ -113,10 +86,6 @@ const Sidebar = ({
 
 const mapStateToProps = (state) => ({
   filters: state.filters,
-  ebolaData: state.ebola.ebolaData.data,
-  ebolaDataCombined: state.ebola.ebolaDataCombined.data,
-  covidCaseCountData: state.covid.caseCounts.data,
-  covidDeathCountData: state.covid.deathCounts.data,
 });
 
 const mapDispatchToProps = (dispatch) =>
